capital-finder/requirements.txt
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
faiss-cpu==1.8.0
streamlit==1.38.0
python-dotenv==1.0.1
tiktoken==0.7.0
